refactor: move EthereumAPI functions into Execution/Consensus API
